Fixes for newlib on new kernel

+++ newlib-redox-backend/common.c
@@ -3,14 +3,14 @@
 extern int errno;
 
 #define demux(a) { \
-    if(a >= (uint)(-4096)){ \
+    if(a >= (uint64_t)(-4096)){ \
         errno = -(int)a; \
-        a = (uint)-1; \
+        a = (uint64_t)-1; \
     } \
-    return (int)a; \
+    return (int64_t)a; \
 }
 
-int syscall0(uint a){
+int64_t syscall0(uint64_t a){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a)
@@ -19,7 +19,7 @@ int syscall0(uint a){
     demux(a)
 }
 
-int syscall1(uint a, uint b){
+int64_t syscall1(uint64_t a, uint64_t b){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a), "b"(b)
@@ -28,7 +28,7 @@ int syscall1(uint a, uint b){
     demux(a)
 }
 
-int syscall2(uint a, uint b, uint c){
+int64_t syscall2(uint64_t a, uint64_t b, uint64_t c){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a), "b"(b), "c"(c)
@@ -37,7 +37,7 @@ int syscall2(uint a, uint b, uint c){
     demux(a)
 }
 
-int syscall3(uint a, uint b, uint c, uint d){
+int64_t syscall3(uint64_t a, uint64_t b, uint64_t c, uint64_t d){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a), "b"(b), "c"(c), "d"(d)
@@ -46,7 +46,7 @@ int syscall3(uint a, uint b, uint c, uint d){
     demux(a)
 }
 
-int syscall4(uint a, uint b, uint c, uint d, uint e){
+int64_t syscall4(uint64_t a, uint64_t b, uint64_t c, uint64_t d, uint64_t e){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a), "b"(b), "c"(c), "d"(d), "S"(e)
@@ -55,7 +55,7 @@ int syscall4(uint a, uint b, uint c, uint d, uint e){
     demux(a)
 }
 
-int syscall5(uint a, uint b, uint c, uint d, uint e, uint f){
+int64_t syscall5(uint64_t a, uint64_t b, uint64_t c, uint64_t d, uint64_t e, uint64_t f){
     asm volatile("int $0x80"
         : "=a"(a)
         : "a"(a), "b"(b), "c"(c), "d"(d), "S"(e), "D"(f)
